refactor: DiaryDetailScreen 댓글 영역 분리 및 카운트 중복 제거 (#137)
1124줄 단일 파일에 뭉쳐 있던 댓글 UI/로직을 분리하고 총 댓글 수 계산
중복을 제거한다. 동작 보존(1124 → 641줄).

- components/DiaryCommentSection: DiaryCommentSection + DiaryMobileCommentBar
  를 별도 파일로 분리
- hooks/useCommentTree: 댓글/대댓글 조회 → CommentNode 트리 + 위임/차단/
  카운트용 파생 구조 생성 로직 추출
- hooks/useCommentThreadDelegation: 아바타 클릭 이동 + 삭제행 클릭 차단
  이벤트 위임 추출
- utils/commentCount: totalCommentCount 계산을 순수 함수 getDiaryCommentTotal
  로 통합(DiaryCommentSection·DiaryDetailView 2중복 제거) + 특성화 테스트

COMMENT_LIST_PARAMS/REPLIES_MAP_PARAMS 는 useCommentTree 에서 export 해
DiaryDetailView 와 공유한다.
